+++ src/AdaskoTheBeAsT.Interop.Execution/PooledVoidExecutionWorkItem.cs
@@ -1,4 +1,3 @@
-#if NET8_0_OR_GREATER
 using System.Collections.Concurrent;
 using System.Threading.Tasks.Sources;
 
@@ -142,4 +141,3 @@ internal sealed class PooledVoidExecutionWorkItem<TSession>
         Pool.Enqueue(this);
     }
 }
-#endif