+++ test/unit/AdaskoTheBeAsT.Interop.Execution.DependencyInjection.Test/DiSecondTestSession.cs
@@ -1,0 +1,11 @@
+namespace AdaskoTheBeAsT.Interop.Execution.DependencyInjection.Test;
+
+public sealed class DiSecondTestSession
+{
+    public DiSecondTestSession(int sessionId)
+    {
+        SessionId = sessionId;
+    }
+
+    public int SessionId { get; }
+}